// src/shared/constants/enums.ts
export enum CurrentLivingStatus {
  Parents,
  Father,
  Mother,
  Guardian,
}

export const applyForCurrentLivingStatusOptions = Object.keys(CurrentLivingStatus)
  .filter((key) => isNaN(Number(key)))
  .map((key, value) => ({
    label: key,
    value: value.toString(),
  }));